Renamed Generator to CodeGenerator namespace #26
