+++ spec/CodeGenerator/PhpParser/Methods/ConstructorGeneratorSpec.php
@@ -1,20 +1,21 @@
 <?php
 
 declare (strict_types = 1);
-namespace spec\NullDev\Skeleton\Generator\PhpParser\Methods;
+namespace spec\NullDev\Skeleton\CodeGenerator\PhpParser\Methods;
 
 use PhpParser\BuilderFactory;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
-class GetterGeneratorSpec extends ObjectBehavior
+class ConstructorGeneratorSpec extends ObjectBehavior
 {
     public function let(BuilderFactory $builderFactory)
     {
         $this->beConstructedWith($builderFactory);
     }
+
     public function it_is_initializable()
     {
-        $this->shouldHaveType('NullDev\Skeleton\Generator\PhpParser\Methods\GetterGenerator');
+        $this->shouldHaveType('NullDev\Skeleton\CodeGenerator\PhpParser\Methods\ConstructorGenerator');
     }
 }
