+++ spec/CodeGenerator/PhpParser/Methods/ToStringGeneratorSpec.php
@@ -1,13 +1,13 @@
 <?php
 
 declare (strict_types = 1);
-namespace spec\NullDev\Skeleton\Generator\PhpParser\Methods;
+namespace spec\NullDev\Skeleton\CodeGenerator\PhpParser\Methods;
 
 use PhpParser\BuilderFactory;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
-class DeserializeGeneratorSpec extends ObjectBehavior
+class ToStringGeneratorSpec extends ObjectBehavior
 {
     public function let(BuilderFactory $builderFactory)
     {
@@ -15,6 +15,6 @@ class DeserializeGeneratorSpec extends ObjectBehavior
     }
     public function it_is_initializable()
     {
-        $this->shouldHaveType('NullDev\Skeleton\Generator\PhpParser\Methods\DeserializeGenerator');
+        $this->shouldHaveType('NullDev\Skeleton\CodeGenerator\PhpParser\Methods\ToStringGenerator');
     }
 }
