+++ spec/CodeGenerator/PhpParser/Methods/UuidCreateGeneratorSpec.php
@@ -1,21 +1,20 @@
 <?php
 
 declare (strict_types = 1);
-namespace spec\NullDev\Skeleton\Generator\PhpParser;
+namespace spec\NullDev\Skeleton\CodeGenerator\PhpParser\Methods;
 
 use PhpParser\BuilderFactory;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
-class ClassGeneratorSpec extends ObjectBehavior
+class UuidCreateGeneratorSpec extends ObjectBehavior
 {
     public function let(BuilderFactory $builderFactory)
     {
         $this->beConstructedWith($builderFactory);
     }
-
     public function it_is_initializable()
     {
-        $this->shouldHaveType('NullDev\Skeleton\Generator\PhpParser\ClassGenerator');
+        $this->shouldHaveType('NullDev\Skeleton\CodeGenerator\PhpParser\Methods\UuidCreateGenerator');
     }
 }
