+++ spec/CodeGenerator/PhpParserGeneratorFactorySpec.php
@@ -1,7 +1,7 @@
 <?php
 
 declare (strict_types = 1);
-namespace spec\NullDev\Skeleton\Generator;
+namespace spec\NullDev\Skeleton\CodeGenerator;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
@@ -10,6 +10,6 @@ class PhpParserGeneratorFactorySpec extends ObjectBehavior
 {
     public function it_is_initializable()
     {
-        $this->shouldHaveType('NullDev\Skeleton\Generator\PhpParserGeneratorFactory');
+        $this->shouldHaveType('NullDev\Skeleton\CodeGenerator\PhpParserGeneratorFactory');
     }
 }
