+++ spec/CodeGenerator/PhpParserGeneratorSpec.php
@@ -1,10 +1,10 @@
 <?php
 
 declare (strict_types = 1);
-namespace spec\NullDev\Skeleton\Generator;
+namespace spec\NullDev\Skeleton\CodeGenerator;
 
-use NullDev\Skeleton\Generator\PhpParser\ClassGenerator;
-use NullDev\Skeleton\Generator\PhpParser\MethodFactory;
+use NullDev\Skeleton\CodeGenerator\PhpParser\ClassGenerator;
+use NullDev\Skeleton\CodeGenerator\PhpParser\MethodFactory;
 use PhpParser\BuilderFactory;
 use PhpParser\PrettyPrinterAbstract;
 use PhpSpec\ObjectBehavior;
@@ -23,6 +23,6 @@ class PhpParserGeneratorSpec extends ObjectBehavior
 
     public function it_is_initializable()
     {
-        $this->shouldHaveType('NullDev\Skeleton\Generator\PhpParserGenerator');
+        $this->shouldHaveType('NullDev\Skeleton\CodeGenerator\PhpParserGenerator');
     }
 }
