+++ src/CodeGenerator/PhpParser/Methods/UuidCreateGenerator.php
@@ -1,7 +1,7 @@
 <?php
 
 declare (strict_types = 1);
-namespace NullDev\Skeleton\Generator\PhpParser\Methods;
+namespace NullDev\Skeleton\CodeGenerator\PhpParser\Methods;
 
 use NullDev\Skeleton\Definition\PHP\Methods\UuidCreateMethod;
 use PhpParser\BuilderFactory;
