+++ tests/CodeGenerator/PhpParserGeneratorTest.php
@@ -4,6 +4,15 @@ declare (strict_types = 1);
 namespace tests\NullDev\Skeleton\Output\PHP;
 
 use Mockery as m;
+use NullDev\Skeleton\CodeGenerator\PhpParser\ClassGenerator;
+use NullDev\Skeleton\CodeGenerator\PhpParser\MethodFactory;
+use NullDev\Skeleton\CodeGenerator\PhpParser\Methods\ConstructorGenerator;
+use NullDev\Skeleton\CodeGenerator\PhpParser\Methods\DeserializeGenerator;
+use NullDev\Skeleton\CodeGenerator\PhpParser\Methods\GetterGenerator;
+use NullDev\Skeleton\CodeGenerator\PhpParser\Methods\SerializeGenerator;
+use NullDev\Skeleton\CodeGenerator\PhpParser\Methods\ToStringGenerator;
+use NullDev\Skeleton\CodeGenerator\PhpParser\Methods\UuidCreateGenerator;
+use NullDev\Skeleton\CodeGenerator\PhpParserGenerator;
 use NullDev\Skeleton\Definition\PHP\DefinitionFactory;
 use NullDev\Skeleton\Definition\PHP\Methods\ConstructorMethod;
 use NullDev\Skeleton\Definition\PHP\Parameter;
@@ -13,15 +22,6 @@ use NullDev\Skeleton\Definition\PHP\Types\TraitType;
 use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\ArrayType;
 use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\IntType;
 use NullDev\Skeleton\Definition\PHP\Types\TypeDeclaration\StringType;
-use NullDev\Skeleton\Generator\PhpParser\ClassGenerator;
-use NullDev\Skeleton\Generator\PhpParser\MethodFactory;
-use NullDev\Skeleton\Generator\PhpParser\Methods\ConstructorGenerator;
-use NullDev\Skeleton\Generator\PhpParser\Methods\DeserializeGenerator;
-use NullDev\Skeleton\Generator\PhpParser\Methods\GetterGenerator;
-use NullDev\Skeleton\Generator\PhpParser\Methods\SerializeGenerator;
-use NullDev\Skeleton\Generator\PhpParser\Methods\ToStringGenerator;
-use NullDev\Skeleton\Generator\PhpParser\Methods\UuidCreateGenerator;
-use NullDev\Skeleton\Generator\PhpParserGenerator;
 use NullDev\Skeleton\Source\ImprovedClassSource;
 use NullDev\Skeleton\Source\SourceFactory;
 use NullDev\Skeleton\SourceFactory\Broadway\CommandSourceFactory;
