+++ tests/CodeGenerator/sample-files/broadway-event.output
