+++ tests/CodeGenerator/sample-files/class-with-all-1-classless-param.output
