+++ tests/CodeGenerator/sample-files/class-with-all-1-param-with-getters.output
