+++ tests/CodeGenerator/sample-files/class-with-all-1-scalartypes-param.output
