+++ tests/CodeGenerator/sample-files/class-with-all-2-param.output
