+++ tests/CodeGenerator/sample-files/class-with-all-3-param.output
