+++ tests/CodeGenerator/sample-files/class-with-all-multi.output
