+++ tests/CodeGenerator/sample-files/class-with-all.output
