+++ tests/CodeGenerator/sample-files/class-with-interface.output
