+++ tests/CodeGenerator/sample-files/class-with-parent.output
