+++ tests/CodeGenerator/sample-files/class-with-trait.output
