+++ tests/CodeGenerator/sample-files/class.output
