+++ tests/CodeGenerator/sample-files/uuid-identifier.output
